Fix autocb for fat arrow functions.
Previously, when generating fat arrow functions, the code
"(autocb) => ..." would get

transformed into an incorrect form:
"(function(_this) { autocb(function(autocb) {..}"}))(this)"

This was due to a bug in handling of the icedHasAutocbFlag when
generating an implicit return of an expression. Code blocks have
an icedHasAutocbFlag that refers to their children, not to their
parent context.