Fix #146
Deferral variables were being given 'var' declarations even when the
variable would have the same name as a parameter in an outer scope,
causing the parameter variable to be overridden.

This is fixed by checking whether the deferral variable already exists
in the current scope before adding it to the scope.