Support for top level Await.
If a top level Await occurs, entire top level Block is wrapped in a
function. This is mostly done to keep the change minimal and yet still
functioning, because it is not needed from JavaScript perspective.t Await
is currently written in a way that it expects a parent function.