Save inline runtime in string form
Inline runtime will now contain stringified copy of itself. This is done
in order to be able to emit the runtime from a browser-based compiler
(or any environment that does not have a filesystem or access to
compiler files).

Removed uglifyjs support as it did not work with ES6 code.

Renamed "build:inline_runtime" task to "build:inline-runtime"