Changed how Iced version is generated
There is now a ICED_PATCH_VERSION field in coffee-script.coffee that
should be used to version the iced patchset itself. The final version of
Iced compiler lands in ICED_VERSION field, that is derived from both VERSION
(which is the mainline CoffeeScript version) and ICED_PATCH_VERSION.

Changed "Generated by" header and '--version' command flag to use
ICED_VERSION. Also '--version' now correctly states that it is
"IcedCoffeeScript", not "CoffeeScript".

Release script was updated to respect these changes.