Add two compilation modes - coffee and iced, to handle `await`
In iced mode, await will be treated as iced await. In coffee mode, await
will be treated as ES6 await. It needed a change in the lexer to emit
lexems differently based on the mode, and it's mostly in the compiler
scaffolding - nodes.coffee was left unchanged apart from one change not
to rely on identifier name when looking for ES6 Await node.